% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module reducer.R
\name{Module.reducer}
\alias{Module.reducer}
\title{Title}
\usage{
Module.reducer(Thresh.Adj.Mat, threshold, nnodes.mod = 5)
}
\arguments{
\item{Thresh.Adj.Mat}{A thresholded square adjecency matrix}

\item{threshold}{the threshold that thresholded the adjecency matrix}

\item{nnodes.mod}{number of nodes in the smallest possible module, default is 5}
}
\value{
the node roles based on the modified module size, and a the network with the outmodule vertices removed
}
\description{
Title
}
