% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nblike.R
\name{nblike}
\alias{nblike}
\title{Title}
\usage{
nblike(theta, n)
}
\arguments{
\item{theta}{Yo no se?}

\item{n}{IDK, Will wrote this one}
}
\value{
Something that I thnk has to do with degree distros
}
\description{
Title
}
