% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node role.R
\name{node.role}
\alias{node.role}
\title{Title}
\usage{
node.role(Thresh.Adj.Mat, threshold, nnodes.mod = 5)
}
\arguments{
\item{Thresh.Adj.Mat}{A thresholded square adjecency matrix}

\item{threshold}{the threshold that thresholded the adjecency matrix}

\item{nnodes.mod}{number of nodes in the smallest possible module, default is 5}
}
\value{
the node roles based on the modified module size, and a the network with the outmodule vertices removed
}
\description{
Title
}
