% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QD network Plotting.R
\name{QD_Network_Plot}
\alias{QD_Network_Plot}
\title{Title}
\usage{
QD_Network_Plot(Tall.Mat, MAINTEXT = "")
}
\arguments{
\item{Tall.Mat}{A thresholded ADJ mat that has been melted into tall format (3 columns, origin, destination and weight of the edge)}

\item{MAINTEXT}{A title for the plot}
}
\value{
The output from \code{\link{plot}} An ugly, but representitive network
}
\description{
Title
}
