% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QD network Plotting.R
\name{QD_Network_Plot_with_Network}
\alias{QD_Network_Plot_with_Network}
\title{Title}
\usage{
QD_Network_Plot_with_Network(Network, MAINTEXT = "")
}
\arguments{
\item{Network}{A Network from iGraph package}

\item{MAINTEXT}{A title for the plot}
}
\value{
The output from \code{\link{plot}} An ugly, but representitive network
}
\description{
Title
}
