% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm get likelyhood.R
\name{rm.get.likelyhood}
\alias{rm.get.likelyhood}
\title{Title}
\usage{
rm.get.likelyhood(rand.mat, nr.thresholds = 101,
  unfold.method = "gaussian", bandwidth = "nrd0",
  nr.fit.points = 101, dist.method = "LL", nr.breaks = 101,
  discard.outliers = TRUE, discard.zeros = FALSE, min.mat.dim = 40,
  max.ev.spacing = 3, interval = NULL, interactive = F,
  smooth.par = 0.5, plot.comp = F, save.fit = FALSE,
  plot.spacing = FALSE, wait.seconds = 0)
}
\arguments{
\item{rand.mat}{A Square correlation matrix, AKA adj. mat}

\item{nr.thresholds}{101}

\item{unfold.method}{Gaussian}

\item{bandwidth}{"nrd0"}

\item{nr.fit.points}{101}

\item{dist.method}{"LL"}

\item{nr.breaks}{101}

\item{discard.outliers}{TRUE}

\item{discard.zeros}{FALSE}

\item{min.mat.dim}{Minimum number of speices for calculations to proceed. Lower for datasets with low diversity}

\item{max.ev.spacing}{3}

\item{interval}{NULL}

\item{interactive}{FALSE}

\item{smooth.par}{0.5}

\item{plot.comp}{FALSE}

\item{save.fit}{FALSE}

\item{plot.spacing}{FALSE}

\item{wait.seconds}{Not important}
}
\value{
The output from \code{\link{cbind}}

The output from \code{\link{print}}
}
\description{
Title
}
