% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R and Z Transform.R
\name{rntransform}
\alias{rntransform}
\title{Title https://www.rdocumentation.org/packages/GenABEL/versions/1.8-0/topics/rntransform}
\usage{
rntransform(formula, data, family = gaussian)
}
\arguments{
\item{formula}{GLM formula for the variable to be transformed, or just the variable}

\item{data}{data.frame or gwaa.data object containing the data}

\item{family}{GLM family}
}
\description{
Title https://www.rdocumentation.org/packages/GenABEL/versions/1.8-0/topics/rntransform
}
