% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecGen.R
\name{Spec.Gen}
\alias{Spec.Gen}
\title{Title}
\usage{
Spec.Gen(Thresholded.Matrix, abun.min = 0.02, upper.site.limit = 0.5,
  lower.site.limit = 0.1)
}
\arguments{
\item{Thresholded.Matrix}{A thresholded matrix, species as columns and samples as rows.}

\item{abun.min}{The minimal local abundance a specialist species must average in across all sites it occurs. Default is 0.02}

\item{upper.site.limit}{The maximal number of sites a specialist can occur in. Must be a proportion, defaults to 0.1}

\item{lower.site.limit}{The minimal number of sites a genralist can occur in. Must be a proportion, defaults to 0.5}
}
\value{
A matrix object with columns "No Sites", :Mean Abiundance" The mean abudance of that species across all sites it occurs,
        "Log abundance", the log of the total abundance of that species across all sites it occurs, and finally Specialist, generalist
        and nominal deliniation. A Generalist occurs in equal or more sites than the upper.site.limit, a specialist occurs in
        fewer or equal to sites than the lower.site.limit, and also must have a mean abundace greater than or equal to the
        abun.min.
        Nominal is all else.
}
\description{
Title
}
