% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R and Z Transform.R
\name{ztransform}
\alias{ztransform}
\title{Title}
\usage{
ztransform(formula, data, family = gaussian)
}
\arguments{
\item{formula}{GLM formula for the variable to be transformed, or just the variable}

\item{data}{data.frame or gwaa.data object containing the data}

\item{family}{GLM Family}
}
\description{
Title
}
