% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_metadata_labels.R
\name{apply_labels}
\alias{apply_labels}
\title{apply_labels}
\usage{
apply_labels(df, type = c("columns", "values"))
}
\arguments{
\item{df}{A dataframe to check.}

\item{type}{Either "columns" or "values". Determines what to label.}
}
\value{
A dataframe (depends on \code{type}).
}
\description{
apply_labels
}
