% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_dictionary.R
\name{llm_dictionary}
\alias{llm_dictionary}
\title{llm_dictionary}
\usage{
llm_dictionary(df, type = c("columns", "values", api_key = NA))
}
\arguments{
\item{df}{A dataframe to check.}

\item{type}{Either "columns" or "values". Determines what kind of discrepancy to identify.}
}
\value{
A tibble or list of discrepancies (depends on \code{type}).
}
\description{
llm_dictionary
}
