% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dictionary.R
\name{use_dictionary}
\alias{use_dictionary}
\title{use_dictionary
Compare a dataframe to the data dictionary}
\usage{
use_dictionary(df, type = c("columns", "values"))
}
\arguments{
\item{df}{A dataframe to check.}

\item{type}{Either "columns" or "values". Determines what kind of discrepancy to identify.}
}
\value{
A tibble or list of discrepancies (depends on \code{type}).
}
\description{
use_dictionary
Compare a dataframe to the data dictionary
}
